fix subpackage management
