+++ src/dMG/core/__init__.py
@@ -1,0 +1,8 @@
+from . import calc, data, post, utils
+
+__all__ = [
+    'calc',
+    'data',
+    'post',
+    'utils',
+]
