+++ src/dMG/core/calc/__init__.py
@@ -1,4 +1,4 @@
-from dMG.core.calc.metrics import Metrics
+from .metrics import Metrics
 
 __all__ = [
     'Metrics',
