+++ src/dMG/core/data/loaders/__init__.py
@@ -1,6 +1,6 @@
-from dMG.core.data.loaders.base import BaseLoader
-from dMG.core.data.loaders.hydro_loader import HydroLoader
-from dMG.core.data.loaders.ms_hydro_loader import MsHydroLoader
+from .base import BaseLoader
+from .hydro_loader import HydroLoader
+from .ms_hydro_loader import MsHydroLoader
 
 __all__ = [
     'BaseLoader',
