+++ src/dMG/core/data/samplers/__init__.py
@@ -1,6 +1,6 @@
-from dMG.core.data.samplers.base import BaseSampler
-from dMG.core.data.samplers.hydro_sampler import HydroSampler
-from dMG.core.data.samplers.ms_hydro_sampler import MsHydroSampler
+from .base import BaseSampler
+from .hydro_sampler import HydroSampler
+from .ms_hydro_sampler import MsHydroSampler
 
 __all__ = [
     'BaseSampler',
