+++ src/dMG/core/data/samplers/hydro_sampler.py
@@ -4,7 +4,7 @@ import numpy as np
 import torch
 from numpy.typing import NDArray
 
-from dMG.core.data import random_index
+from dMG.core.data.data import random_index
 from dMG.core.data.samplers.base import BaseSampler
