+++ src/dMG/models/__init__.py
@@ -1,0 +1,9 @@
+from . import criterion, deltamodel, neural_networks, phy_models
+
+__all__ = [
+    'criterion',
+    'deltamodel',
+    'neural_networks',
+    'phy_models',
+    'ModelHandler',
+]
