+++ src/dMG/models/criterion/__init__.py
@@ -1,12 +1,12 @@
-from dMG.models.criterion.base import BaseCriterion
-from dMG.models.criterion.kge_batch_loss import KgeBatchLoss
-from dMG.models.criterion.kge_norm_batch_loss import KgeNormBatchLoss
-from dMG.models.criterion.mse_loss import MSELoss
-from dMG.models.criterion.nse_batch_loss import NseBatchLoss
-from dMG.models.criterion.nse_sqrt_batch_loss import NseSqrtBatchLoss
-from dMG.models.criterion.range_bound_loss import RangeBoundLoss
-from dMG.models.criterion.rmse_comb_loss import RmseCombLoss
-from dMG.models.criterion.rmse_loss import RmseLoss
+from .base import BaseCriterion
+from .kge_batch_loss import KgeBatchLoss
+from .kge_norm_batch_loss import KgeNormBatchLoss
+from .mse_loss import MSELoss
+from .nse_batch_loss import NseBatchLoss
+from .nse_sqrt_batch_loss import NseSqrtBatchLoss
+from .range_bound_loss import RangeBoundLoss
+from .rmse_comb_loss import RmseCombLoss
+from .rmse_loss import RmseLoss
 
 __all__ = [
     'BaseCriterion',
