+++ src/dMG/models/deltamodel/__init__.py
@@ -1,4 +1,4 @@
-from dMG.models.deltamodel.dpl_model import DplModel as dPLModel
+from .dpl_model import DplModel as dPLModel
 
 __all__ = [
     'dPLModel',
