+++ src/dMG/models/neural_networks/__init__.py
@@ -1,8 +1,8 @@
-from dMG.models.neural_networks.ann import AnnCloseModel
-from dMG.models.neural_networks.cudnn_lstm import CudnnLstmModel
-from dMG.models.neural_networks.lstm import LstmModel
-from dMG.models.neural_networks.lstm_mlp import LstmMlpModel
-from dMG.models.neural_networks.mlp import MlpMulModel
+from .ann import AnnCloseModel
+from .cudnn_lstm import CudnnLstmModel
+from .lstm import LstmModel
+from .lstm_mlp import LstmMlpModel
+from .mlp import MlpMulModel
 
 __all__ = [
     'CudnnLstmModel',
